scrapers/eztv: Use regex instead of BeautifulSoup to get episodes list
BeautifulSoup is **too** slow in a raspberry pi and we are using very basic
things from it. So, as it was simple, this commit just replaces the uses of
BeautifulSoup to get season episodes using regex.

Also, the patch tries to be as simple as possible (to make more easy to catch
the mistakes). The "big" difference in the diff is that a whole BeautifulSoup
node was saved and then access using ".text" or ".href", and now we only save
the text (by using a regex group). So, in the diff lot of ".text" and ".href"
are removed.

Before this commit on my raspberry pi it took 45 seconds to get the episodes for
White Collar Season 5, and with this path it takes 10 seconds.